LayoutTests/imported/w3c/web-platform-tests/xhr/send-conditional-cors-expected.txt

CONSOLE MESSAGE: Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/xhr/resources/conditional.py?tag=0.017259906889160348 due to access control checks.

PASS 304 without appropriate CORS header 
PASS 304 with appropriate CORS header 
